Refactor code
- Change naming convention of controllers, models, routes file
- Move route file from api to routes directory
